minor helper refactor
-moved getImage to utils
-added basic jsdocs to helpers
