+++ src/edit image at runtime.js
@@ -1,6 +1,8 @@
-/*
-bitsy hack helper - edit image at runtime
-
+/**
+@file edit image at runtime
+@summary API for updating image data at runtime.
+@author Sean S. LeBlanc
+@description
 Adds API for updating sprite, tile, and item data at runtime.
 
 Individual frames of image data in bitsy are 8x8 1-bit 2D arrays in yx order
@@ -17,6 +19,13 @@ e.g. the default player is:
 ]
 */
 import bitsy from "bitsy";
+import {
+	getImage
+} from "./utils.js";
+
+export {
+	getImage
+};
 
 /*
 Args:
@@ -77,20 +86,4 @@ export function setTileData(id, frame, newData) {
 
 export function setItemData(id, frame, newData) {
 	setImageData(id, frame, bitsy.item, newData);
-}
-
-/*
-Helper for getting image by name or id
-
-Args:
-	name: id or name of image to return
-	 map: map of images (e.g. `sprite`, `tile`, `item`)
-
-Returns: the image in the given map with the given name/id
- */
-export function getImage(name, map) {
-	var id = map.hasOwnProperty(name) ? name : Object.keys(map).find(function (e) {
-		return map[e].name == name;
-	});
-	return map[id];
 }
