+++ src/utils.js
@@ -1,3 +1,9 @@
+/**
+@file utils
+@summary miscellaneous bitsy utilities
+@author Sean S. LeBlanc
+*/
+
 /*helper used to inject code into script tags based on a search string*/
 export function inject(searchString, codeToInject) {
 	// find the relevant script tag
@@ -38,4 +44,20 @@ export function expose(target) {
 	code += "this.get = function(name) {return eval(name);};";
 	code += "this.set = function(name, value) {eval(name+'=value');};";
 	return eval("[" + code + "}]")[0];
+}
+
+/*
+Helper for getting image by name or id
+
+Args:
+	name: id or name of image to return
+	 map: map of images (e.g. `sprite`, `tile`, `item`)
+
+Returns: the image in the given map with the given name/id
+ */
+export function getImage(name, map) {
+	var id = map.hasOwnProperty(name) ? name : Object.keys(map).find(function (e) {
+		return map[e].name == name;
+	});
+	return map[id];
 }
